+++ src/app/app.component.ts
@@ -1,8 +1,6 @@
-import {Component, OnInit} from '@angular/core';
+import {AfterViewInit, Component, ElementRef, OnInit, ViewChild} from '@angular/core';
 import {ProductRepository} from "./product/services/product-repository";
-import {CartService} from "./cart/components/services/cart-service";
-import {ProductModel} from "./product/models/product-model";
-import {CartItemModel} from "./cart/components/models/cart-item-model";
+import {CartService} from "./cart/services/cart-service";
 
 @Component({
   selector: 'app-root',
@@ -13,28 +11,16 @@ import {CartItemModel} from "./cart/components/models/cart-item-model";
     CartService,
   ],
 })
-export class AppComponent implements OnInit {
-  public products: ProductModel[] = [];
+export class AppComponent implements OnInit, AfterViewInit {
+  @ViewChild('appTitle') public title!: ElementRef<HTMLHeadingElement>;
 
-  get cartItems(): CartItemModel[] {
-    return this.cartService.items;
+  constructor() {
   }
 
-  get cartTotal(): number {
-    return this.cartService.total;
+  public ngOnInit(): void {
   }
 
-  constructor(
-    private productRepository: ProductRepository,
-    private cartService: CartService,
-  ) {
-  }
-
-  async ngOnInit(): Promise<void> {
-    this.products = await this.productRepository.getProducts();
-  }
-
-  public onAddItemToCart(item: ProductModel): void {
-    this.cartService.addItemToCart(item);
+  public ngAfterViewInit(): void {
+    this.title.nativeElement.innerText = 'Test title';
   }
 }