+++ src/app/shared/directives/highlight-directive.directive.ts
@@ -1,0 +1,28 @@
+import {Directive, ElementRef, HostBinding, HostListener, Input, Renderer2} from '@angular/core';
+
+@Directive({
+  selector: '[appHighlightDirective]'
+})
+export class HighlightDirective {
+  @Input() public defaultColor: string = '#FFFFFF';
+  @Input() public highlightColor: string = '#CCCCCC';
+
+  private currentColor: string = '#FFFFFF';
+
+  @HostBinding('style.background-color') get getBackgroundColor() {
+    return this.currentColor;
+  }
+
+  @HostListener('mouseover') onMouseOver() {
+    console.log(222)
+    this.currentColor = this.highlightColor;
+  }
+
+  @HostListener('mouseout') onMouseOut() {
+    this.currentColor = this.defaultColor;
+  }
+
+  constructor() {
+  }
+
+}